+++ lib/collaboration/coherence/invitation.ex
@@ -5,9 +5,6 @@ defmodule Collaboration.Coherence.Invitation do
   use Ecto.Schema
   import Ecto.Changeset
 
-  @primary_key {:id, :binary_id, autogenerate: true}
-  @foreign_key_type :binary_id
-
   schema "invitations" do
     field(:name, :string)
     field(:email, :string)