+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### 2.3.2
  * removed login via facebook or google placeholders
  * separated participants from users in admin list view
+ * allows for eight conditions now
+ * switched from binary ids to int ids for better performance
+ * removed unlock, active and confirmation mechanics (unnecessary)
 
 ### 2.3.1
  * fixed port number in email links